requirements.txt
flask
python-binance
nano requirements.txt
